+++ src/main/java/net/semperidem/fishingclub/fisher/managers/ProgressionManager.java
@@ -190,7 +190,7 @@ public class ProgressionManager extends DataManager {
         return this.knownTradeSecrets.get(tradeSecret.name()).level();
     }
 
-    public int tradeSecretValue(TradeSecret tradeSecret) {
+    public float tradeSecretValue(TradeSecret tradeSecret) {
         if (!this.knowsTradeSecret(tradeSecret)) {
             return 0;
         }
